Lab-6-storage/requirements.txt
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
clickclick==20.10.2
connexion==2.14.2
Flask==2.2.5
greenlet==2.0.2
idna==3.4
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kazoo==2.5.0
MarkupSafe==2.1.3
mysql-connector-python==8.1.0
packaging==23.1
protobuf==4.21.12
pykafka==2.8.0
PyMySQL==1.1.0
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.3
six==1.16.0
SQLAlchemy==2.0.21
swagger-ui-bundle==0.0.9
tabulate==0.9.0
typing_extensions==4.8.0
urllib3==2.0.5
Werkzeug==2.2.3
